Core: Move Vector2i to its own `vector2i.h` header
Also reduce interdependencies and clean up a bit.
