+++ core/math/rect2.h
@@ -31,6 +31,7 @@
 #ifndef RECT2_H
 #define RECT2_H
 
+#include "core/error/error_macros.h"
 #include "core/math/vector2.h"
 
 class String;
