+++ core/variant/variant.h
@@ -44,6 +44,7 @@
 #include "core/math/transform_2d.h"
 #include "core/math/transform_3d.h"
 #include "core/math/vector2.h"
+#include "core/math/vector2i.h"
 #include "core/math/vector3.h"
 #include "core/math/vector3i.h"
 #include "core/object/object_id.h"
