+++ editor/editor_atlas_packer.h
@@ -31,10 +31,11 @@
 #ifndef EDITOR_ATLAS_PACKER_H
 #define EDITOR_ATLAS_PACKER_H
 
-#include "core/math/vector2.h"
-
 #include "core/templates/vector.h"
 #include "scene/resources/bit_map.h"
+
+struct Vector2;
+struct Vector2i;
 
 class EditorAtlasPacker {
 public:
