+++ modules/gdnative/gdnative/rect2.cpp
@@ -32,6 +32,7 @@
 
 #include "core/math/rect2.h"
 #include "core/math/rect2i.h"
+#include "core/os/memory.h"
 
 static_assert(sizeof(godot_rect2) == sizeof(Rect2), "Rect2 size mismatch");
 static_assert(sizeof(godot_rect2i) == sizeof(Rect2i), "Rect2i size mismatch");
