+++ modules/gdnative/gdnative/transform2d.cpp
@@ -31,6 +31,7 @@
 #include "gdnative/transform2d.h"
 
 #include "core/math/transform_2d.h"
+#include "core/os/memory.h"
 
 static_assert(sizeof(godot_transform2d) == sizeof(Transform2D), "Transform2D size mismatch");
