+++ modules/gdnative/gdnative/vector2.cpp
@@ -31,6 +31,8 @@
 #include "gdnative/vector2.h"
 
 #include "core/math/vector2.h"
+#include "core/math/vector2i.h"
+#include "core/os/memory.h"
 
 static_assert(sizeof(godot_vector2) == sizeof(Vector2), "Vector2 size mismatch");
 static_assert(sizeof(godot_vector2i) == sizeof(Vector2i), "Vector2i size mismatch");
