+++ tests/core/math/test_vector2i.h
@@ -32,6 +32,7 @@
 #define TEST_VECTOR2I_H
 
 #include "core/math/vector2.h"
+#include "core/math/vector2i.h"
 #include "tests/test_macros.h"
 
 namespace TestVector2i {
